df_imoveis_urls/requirements.txt
python-dotenv 
supabase == 2.17.0
Scrapy == 2.13.3
pandas == 2.3.1
numpy == 2.3.2
google_api == 0.1.12
pathlib == 1.0.1